+++ src/Commands/Projects/DeleteGit.php
@@ -1,0 +1,42 @@
+<?php
+
+namespace Sven\ForgeCLI\Commands\Projects;
+
+use Sven\ForgeCLI\Commands\BaseCommand;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Question\ConfirmationQuestion;
+
+class DeleteGit extends BaseCommand
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function configure()
+    {
+        $this->setName('delete:git')
+            ->addArgument('server', InputArgument::REQUIRED, 'The id of the server the site is on.')
+            ->addArgument('site', InputArgument::REQUIRED, 'The id of the site to delete the git project from.')
+            ->setDescription('Delete a git project from a site.');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function execute(InputInterface $input, OutputInterface $output)
+    {
+        $site = $input->getArgument('site');
+
+        $helper = $this->getHelper('question');
+        $question = new ConfirmationQuestion('Are you sure you want to delete the git project from the site with id "'.$site.'"?', false);
+
+        if (! $helper->ask($input, $output, $question)) {
+            $output->writeln('<info>Ok, aborting. Your git project is safe.</info>');
+
+            return;
+        }
+
+        $this->forge->destroySiteGitRepository($input->getArgument('server'), $site);
+    }
+}