+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `install:git` command
+- `delete:git` command
+- `install:wordpress` command
+- `delete:wordpress` command
 
 ## [0.2.0] - 2017-06-23
 ### Added